backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
httpx==0.26.0
requests==2.31.0

# AWS S3
boto3==1.34.34

# Image Processing
Pillow==10.2.0

# Async Support
aiofiles==23.2.1

# CORS
fastapi-cors==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0
